moved the  custom components to appropriate place
